requirements.txt
# UI
customtkinter>=5.2.0
Pillow>=10.0.0

# Local AI
ollama>=0.4.0

# Browser Automation
playwright>=1.40.0

# Desktop Automation
pyautogui>=0.9.54

# Scheduling
apscheduler>=3.10.0

# Utilities
pydantic>=2.5.0

# Voice (optional — Nexus works without these, voice just won't activate)
sounddevice>=0.4.6
numpy>=1.24.0
openai-whisper>=20240930
pyttsx3>=2.90

# Screen Awareness (optional)
mss>=9.0.0

# Memory fuzzy search (optional)
rapidfuzz>=3.5.0

# System Monitor — real-time CPU, RAM, disk stats
psutil>=5.9.0

# Website Auditor + Weather + Uptime Monitor + Telegram
requests>=2.31.0
beautifulsoup4>=4.12.0

# Web Remote (phone browser control)
flask>=3.0.0

# CAD Engine — AI-driven 3D part design (install separately, large dependency)
# pip install cadquery
# cadquery>=2.4.0

# Proposal Writer — PDF generation
fpdf2>=2.7.0

# Hotkey Daemon — global keyboard shortcuts
keyboard>=0.13.5

# Vision AI — image analysis with LLaVA (Pillow already pulled by CustomTkinter)
# Pillow already in requirements above

# Meeting Notes — audio transcription (whisper already in requirements above)
# openai-whisper already in requirements above

# Auto Documenter / Code Writer / other new plugins use existing deps only

# ── New Plugin Dependencies (v2.2) ─────────────────────────────────────────

# Clipboard AI
pyperclip>=1.8.2

# News Digest — RSS feed parsing
feedparser>=6.0.8

# PDF Reader — try fitz first (PyMuPDF), fall back to pypdf
PyMuPDF>=1.24.0
# pypdf>=4.0.0   # fallback if PyMuPDF too large

# QR Generator
qrcode[pil]>=7.4.2

# Password Vault — encryption
cryptography>=42.0.0

# Screen Recorder — video encoding
imageio>=2.33.0
imageio-ffmpeg>=0.4.9

# Language Coach / Dream Journal / Research — use Ollama (already in deps)

# Local RAG — optional semantic embeddings (huge download, optional)
# sentence-transformers>=2.7.0

# Competitor Tracker — uses requests+bs4 (already in deps)
# All other new plugins use existing dependencies only
